Release 6.2.15 - IMAP sequence sets, and every file has an owner
IMAP sequence sets now follow RFC 3501. "*" was only recognised as the end
of a range and read as zero anywhere else, and descending ranges were never
normalised, so "FETCH *" silently returned nothing, "UID STORE *:* +FLAGS
(\Deleted)" flagged the whole mailbox rather than the newest message, and
"UID EXPUNGE *" deleted every \Deleted message. Fixed in all four
sequence-set parsers - the FETCH/STORE/COPY/MOVE path, SEARCH, UID EXPUNGE
and the QRESYNC VANISHED path - with twelve regression tests that fail
against 6.2.14.

Files the server writes are now accounted for on every path:

  - A failed restore emptied the live data directory before checking that
    the backup contained a message store to put back, then threw and left
    the only remaining copy of the mail in an unreferenced temp folder.
  - A message file that could not be read stalled each outbound attempt
    for up to ten minutes. It now fails immediately, and permanently only
    when the file has genuinely gone; a file merely held open elsewhere
    stays queued rather than being bounced.
  - Rejected POP3 RETR, downloaded messages with no local recipient,
    undeliverable bounces and failed header rewrites each left a file
    behind.
  - The account cache had no size limit, because a duplicated line gave
    the domain cache its cap twice.

From RvdH's forum feedback: BOOST_USE_WINAPI_VERSION now matches
_WIN32_WINNT, upstream's general EOF rule replaces our SpamAssassin-port
special case, and upstream PR #530 (DKIM header-name case under simple
canonicalization, plus the MIME encoding overflow) is ported.

Adopting the EOF rule invalidated an assumption in our BDAT reads:
transfer_exactly only guarantees the full octet count while any error ends
the session. A truncated chunk was padded with NUL bytes and delivered as
complete. Reads now take only what arrived, and no read is re-armed after
end-of-stream.

Also: ManageSieve disconnects and auto-bans after three failed logons,
STATUS RECENT is counted per folder rather than for the selected one, and
the SMTP DATA path logs enough under debug logging to tell a stalled
reception from a stalled accept stage (groundwork for #18, still open).

Control Panel: new Authentication, Administrative access, DNS resolver and
Web services pages cover settings that had no GUI at all; archiving,
BackupMessagesDBOnly and the message-store consistency result are now
visible; four labels that described the wrong thing are corrected.

Full regression suite: 1040 of 1040 passing, zero failures.

+++ hmailserver/source/Server/Common/Mime/MimeCode.h
@@ -291,7 +291,9 @@ namespace HM
 	   void UnfoldField(string& strField) const;
 	   int SelectEncoding(size_t nLength, int nNonAsciiChars) const
 	   {
-         size_t nQEncodeSize = nLength + nNonAsciiChars * 2;
+         // Widen before multiplying: the int product overflowed for very long
+         // header values. (From upstream hMailServer, PR #530.)
+         size_t nQEncodeSize = nLength + ((size_t) nNonAsciiChars) * 2;
          size_t nBEncodeSize = (nLength + 2) / 3 * 4;
 		   return (nQEncodeSize <= nBEncodeSize || ((size_t) nNonAsciiChars)*5 <= nLength) ? 'Q' : 'B';
 	   }
